add pgsqlQuoteIdentifier method
this is used when you need to quote a name of an identifier. While the
PDO::quote() is actually not really needed in case of prepared
statements, a method for quoting identifiers definitely is whenever you
need to dynamically form a query and you need to select user-provided
columns.

Quoting identifiers is as difficult to get right as quoting values, with
the exact same dangers if you don't do it right, so this should really
be a welcome addition to the set of functions.

Yes. This should maybe go into PDO itself, but this now adds is for
postgres at least